requirements.txt
selenium>=4.20.0
webdriver-manager>=4.0.0
pytest>=8.2.0
pytest-xdist>=3.6.1
allure-pytest>=2.13.2
python-dotenv>=1.0.1